requirements.txt
alembic
cryptography>=1.0.1
eralchemy
Flask>=0.10.1
Flask-Login>=0.2.11
Flask-JWT>=0.3.2
Flask-SQLAlchemy>=2.0
Flask-WTF>0.9.5
flask-restful>=0.2.12
flask-security==3.0.0
flasgger>=0.9.2
marshmallow-sqlalchemy
hashids==1.1.0
itsdangerous>=0.24
Jinja2>=2.7.3
MarkupSafe==0.23
passlib>=1.6.2
psycopg2>=2.6.1
py-bcrypt>=0.4
pycrypto>=2.6.1
pytz>=2015.4
PyJWT>=1.3.0
python-jose-cryptodome>=1.3.2
requests
SQLAlchemy>=1.0.4
Werkzeug>=0.10.4
wheel==0.24.0
wscodec
WTForms>=2.0.2
